fix: cast index to u16 in load_from_vec method to ensure correct address calculation

+++ src/memory.rs
@@ -38,7 +38,7 @@ pub trait Addressable {
 
     fn load_from_vec(&mut self, from: &[u8], addr : u16) -> bool {
         for (i,b) in from.iter().enumerate() {
-            if !self.write(addr+i,*b) {
+            if !self.write(addr+(i as u16),*b) {
                 return false;
             }
         };
